+++ diskover_connections.py
@@ -53,4 +53,8 @@ def connect_to_redis():
     from diskover import config
     global redis_conn
     redis_conn = Redis(host=config['redis_host'], port=config['redis_port'],
-                       password=config['redis_password'], db=config['redis_db'])
+                       password=config['redis_password'], db=config['redis_db'],
+                       retry_on_timeout=config['redis_retry_on_timeout'],
+                       socket_keepalive=config['redis_socket_keepalive'],
+                       socket_connect_timeout=config['redis_socket_connect_timeout'],
+                       socket_timeout=config['redis_socket_timeout'])